refactor: Bookmark → UserGroup 으로 테이블명 변경 (#96)
