+++ src/main/java/com/beside/archivist/entity/usergroup/UserGroup.java
@@ -1,4 +1,4 @@
-package com.beside.archivist.entity.bookmark;
+package com.beside.archivist.entity.usergroup;
 
 import com.beside.archivist.entity.group.Group;
 import com.beside.archivist.entity.users.User;
@@ -10,9 +10,9 @@ import lombok.NoArgsConstructor;
 
 @Table @Entity @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Bookmark {
+public class UserGroup {
     @Id
-    @Column(name = "bookmark_id")
+    @Column(name = "user_group_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
@@ -27,7 +27,7 @@ public class Bookmark {
     private Group groups;
 
     @Builder
-    public Bookmark(boolean isOwner, User users, Group groups) {
+    public UserGroup(boolean isOwner, User users, Group groups) {
         this.isOwner = isOwner;
         this.users = users;
         this.groups = groups;
