+++ src/main/java/com/beside/archivist/entity/users/User.java
@@ -1,10 +1,8 @@
 package com.beside.archivist.entity.users;
 
 
-import com.beside.archivist.dto.users.UserDto;
-import com.beside.archivist.entity.BaseEntity;
 import com.beside.archivist.entity.BaseTimeEntity;
-import com.beside.archivist.entity.bookmark.Bookmark;
+import com.beside.archivist.entity.usergroup.UserGroup;
 import com.beside.archivist.entity.link.Link;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
@@ -42,7 +40,7 @@ public class User extends BaseTimeEntity {
     private List<Link> links = new ArrayList<>();
 
     @OneToMany(mappedBy = "users", fetch = FetchType.LAZY, cascade = CascadeType.ALL, orphanRemoval = true)
-    private List<Bookmark> bookmarks = new ArrayList<>();
+    private List<UserGroup> userGroups = new ArrayList<>();
 
     @Builder
     public User(String email, String password, String nickname, List<Category> categories, UserImg userImg) {
@@ -62,7 +60,7 @@ public class User extends BaseTimeEntity {
     public void addLink(Link l){
         this.links.add(l);
     }
-    public void addBookmark(Bookmark b) {
-        this.bookmarks.add(b);
+    public void addUserGroup(UserGroup userGroup) {
+        this.userGroups.add(userGroup);
     }
 }
