+++ src/main/java/com/beside/archivist/service/usergroup/UserGroupService.java
@@ -1,0 +1,8 @@
+package com.beside.archivist.service.usergroup;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public interface UserGroupService {
+    void saveUserGroup(Long groupId);
+}
